mail funcionando

+++ src/dao/mongo/user.mongo.js
@@ -69,13 +69,15 @@ class UserManagerMongo {
             </div>`
         
             users.map( async(u)=>{
-                const ultimaConexion = moment(u.last_connection)
-                const fechaActual = moment(Date.now())
-                const tiempoInactividad = fechaActual.diff(ultimaConexion,'d')
-
-                if(tiempoInactividad > 2){
-                    await sendMail(u.email, subject, html)
-                    return await userModel.deleteOne({_id: u._id})
+                if(u.last_connection){
+                    const ultimaConexion = moment(u.last_connection)
+                    const fechaActual = moment(Date.now())
+                    const tiempoInactividad = fechaActual.diff(ultimaConexion,'d')
+    
+                    if(tiempoInactividad > 2){
+                        await sendMail(u.email, subject, html)
+                        return await userModel.deleteOne({_id: u._id})
+                    }
                 }
             })
         } catch (error) {
